wandu lite first commit

+++ database/migrations/2020_09_11_113631_create_shop_storages_table.php
@@ -22,7 +22,7 @@ class CreateShopStoragesTable extends Migration
             $table->integer("user_id");
             $table->date("date");
             $table->time("time");
-            $table->integer("quantity");
+            $table->float("quantity", 8, 2);
             $table->dateTime("deleted_at")->nullable();
             $table->timestamps();
         });
